+++ ReleaseNotes.md
@@ -1,6 +1,7 @@
 # Release notes of Ampersand
 
 ## current development 
+  * Work on meatgrinder (still experimental!)
   * Bug fix: Issue with SQL query [Issue #152](https://github.com/AmpersandTarski/Ampersand/issues/152)
   * Bug fix: minor issue with SQL query [Issue #436](https://github.com/AmpersandTarski/Ampersand/issues/436)
   * Bug fix: Nontermination of functional document generator. [Issue #231](https://github.com/AmpersandTarski/Ampersand/issues/231)